+++ .eslintrc.js
@@ -2,13 +2,22 @@ module.exports = {
   root: true,
   extends: [
     'airbnb-base',
-    'plugin:jest/all'
+    'plugin:jest/all',
+    'plugin:sonarjs/recommended',
   ],
-  plugins: [],
+  plugins: [
+    'sonarjs',
+  ],
   env: {
     node: true,
   },
   rules: {
+    'sonarjs/cognitive-complexity': 1,
+    'sonarjs/no-collapsible-if': 0,
+    'sonarjs/no-duplicated-branches': 1,
+    'sonarjs/no-identical-functions': 0,
+    'sonarjs/no-duplicate-string': 0,
+    'sonarjs/no-same-line-conditional': 0,
     'implicit-arrow-linebreak': 0,
     'no-param-reassign': 0,
     'import/no-extraneous-dependencies': [